requirements.txt
discord.py>=2.3.2,<3.0.0
python-dotenv>=1.0.0
yt-dlp>=2023.12.30
spotipy>=2.22.1
PyNaCl>=1.5.0
aiohttp>=3.8.0
matplotlib>=3.5.0